Move the force stop function from the All Trans page to the Detail page.
